requirements.txt
fastapi==0.110.1
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.29
alembic==1.13.1
pydantic==2.6.4
pydantic-settings==2.2.1
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.1.2
psycopg[binary]==3.1.18
faker==24.4.0
python-multipart==0.0.9
